Add solution for problem `C - Unstable String` Link: https://codeforces.com/contest/1535/problem/C Tags: binary search, dp, greedy, implementation, strings, two pointers, *1400 Ref: https://codeforces.com/contest/1535/submission/118429383

+++ codeforces/1535/C.cpp
@@ -18,7 +18,7 @@ void solve() {
 	int n = s.size();
 	long long dp[n + 7];
 	dp[0] = 1;
-	int curr = -1, qcount = 1, qs = 0;
+	long long curr = -1, qcount = 1, qs = 0;
 	if (s[0] == '0') {curr = 1;}
 	if (s[0] == '1') {curr = 0;}
 	if (s[0] == '?') {qs++;}
